requirements.txt
pytesseract
flask
cherrypy
opencv-python
numpy
pillowfight
pypillowfight
nltk